Moved GenericApiResponse to tools
